@@ -6,20 +6,23 @@
     import moment from 'moment'
 
     export default {
+        model: {
+            prop: 'model',
+            event: 'set-value',
+        },
+
         props: {
             model: {
                 required: true,
-                twoWay: true,
-            }
+            },
         },
-
         computed: {
             value: {
                 get() {
-                    return this.model.format('YYYY-MM-DD[T]HH:mm')
+                    return moment(this.model).format('YYYY-MM-DD[T]HH:mm')
                 },
                 set(val) {
-                    this.model = moment(val)
+                    this.$emit('set-value', moment(val))
                 },
             },
         },
