@@ -2,32 +2,51 @@
     <div class="ui form">
         <div class="ui two column stackable grid">
             <div class="column" v-for="option in options">
-                <div class="ui toggle checkbox">
-                    <input :disabled="readOnly" type="checkbox" v-model="model[option.value]">
-                    <label>{{ option.field_name }}</label>
-                </div>
+                <semantic-checkbox type="checkbox"
+                                   v-model="checked[option.value]"
+                                   :label="option.field_name"
+                                   @checkbox-clicked="setValue($event, option)">
+                </semantic-checkbox>
             </div>
         </div>
     </div>
 </template>
 
 <script>
-    import Input from '../../mixins/Input.js'
+    import Input from '../../mixins/Input'
 
     export default {
         mixins: [
             Input,
         ],
+        methods: {
+            setValue(e, val) {
+                this.checked[val.value] = !e
+            },
+        },
 
         computed: {
             options() {
                 return Object.keys(this.field.field_options.select_options).map((value) => {
                     const obj = {
+                        // eslint-disable-next-line
                         value: value.replace(/\-/g, '_'),
                         field_name: this.field.field_options.select_options[value],
                     }
                     return obj
                 })
+            },
+            checked() {
+                return this.model
+            },
+        },
+
+        watch: {
+            checked: {
+                deep: true,
+                handler(val) {
+                    this.$emit('set-value', val)
+                },
             },
         },
     }
