@@ -1,27 +1,32 @@
 <template>
-    <semantic-form-dropdown
-    :placeholder="field.name"
-    :model.sync="model"
-    :options="options"
-    :full_text_search="true"
-    :search="true"
-    :fluid="true"
-    :disabled="readOnly"
+    <semantic-form-dropdown :placeholder="field.name"
+                            :value="value"
+                            :options="options"
+                            :full_text_search="true"
+                            :search="true"
+                            :fluid="true"
+                            :disabled="readOnly"
+                            @dropdown-selected="value = $event"
     ></semantic-form-dropdown>
 </template>
 
 <script>
-    import Input from '../../mixins/Input.js'
+    import Input from '../../mixins/Input'
 
     export default {
+        data() {
+            return {
+                value: this.model,
+            }
+        },
         mixins: [
             Input,
         ],
-
         computed: {
             options() {
                 const options = Object.keys(this.field.field_options.select_options).map((value) => {
                     const obj = {
+                        // eslint-disable-next-line
                         value: value.replace(/\-/g, '_'),
                         name: this.field.field_options.select_options[value],
                     }
@@ -29,6 +34,11 @@
                 })
 
                 return options
+            },
+        },
+        watch: {
+            value() {
+                this.$emit('set-value', this.value)
             },
         },
     }
