@@ -1,26 +1,42 @@
 <template>
-    <select v-model="model" class="ui fluid dropdown" :class="{disabled: readOnly}">
-        <option v-for="option in options" :value="option.value">{{ option.field_name }}</option>
-    </select>
+    <semantic-form-dropdown :placeholder="placeholder"
+                            :value="value"
+                            :options="options"
+                            :fluid="true"
+                            :disabled="readOnly"
+                            @dropdown-selected="value = $event"
+    ></semantic-form-dropdown>
 </template>
 
 <script>
-    import Input from '../../mixins/Input.js'
+    import Input from '../../mixins/Input'
 
     export default {
+        data() {
+            return {
+                value: this.model,
+            }
+        },
         mixins: [
             Input,
         ],
-
         computed: {
             options() {
-                return Object.keys(this.field.field_options.select_options).map((value) => {
+                const options = Object.keys(this.field.field_options.select_options).map((value) => {
                     const obj = {
+                        // eslint-disable-next-line
                         value: value.replace(/\-/g, '_'),
-                        field_name: this.field.field_options.select_options[value],
+                        name: this.field.field_options.select_options[value],
                     }
                     return obj
                 })
+
+                return options
+            },
+        },
+        watch: {
+            value() {
+                this.$emit('set-value', this.value)
             },
         },
     }
