@@ -1,13 +1,16 @@
 module.exports = {
+    model: {
+        prop: 'model',
+        event: 'set-value',
+    },
+
     props: {
         field: {
             required: true,
-            twoWay: true,
         },
 
         model: {
             required: true,
-            twoWay: true,
         },
 
         placeholder: {
@@ -20,6 +23,18 @@ module.exports = {
 
         readOnly: {
             default: false,
+        },
+
+        validation: {
+            required: true,
+        },
+    },
+
+    watch: {
+        value() {
+            if (this.validation.$touch) {
+                this.validation.$touch()
+            }
         },
     },
 }
