@@ -1,13 +1,21 @@
+import Vue from 'vue';
+Vue.config.productionTip = false;
+require('es6-promise').polyfill();
+
+window.jQuery = window.$ = require('jquery')
+
+window.gateway_url = 'gateway.croudcontrol.dev'
+
 // Polyfill fn.bind() for PhantomJS
 /* eslint-disable no-extend-native */
-Function.prototype.bind = require('function-bind')
+Function.prototype.bind = require('function-bind');
 
 // require all test files (files that ends with .spec.js)
-var testsContext = require.context('./specs', true, /\.spec$/)
-testsContext.keys().forEach(testsContext)
+const testsContext = require.context('./specs', true, /\.spec$/);
+testsContext.keys().forEach(testsContext);
 
 // require all src files except main.js for coverage.
 // you can also change this to match only the subset of files that
 // you want coverage for.
-var srcContext = require.context('../../src', true, /^\.\/(?!main(\.js)?$)/)
-srcContext.keys().forEach(srcContext)
+const srcContext = require.context('../../src', true, /^\.\/(?!main(\.js)?$)/);
+srcContext.keys().forEach(srcContext);
